Move card grade related things closer to cards
