+++ src/main/scala/kiteytys/Game.scala
@@ -5,35 +5,12 @@ import java.time.LocalDateTime
 import org.http4s.UrlForm
 import scala.util.Try
 
-object FormParsing {
-  def stringToInt(s: String): Option[Int] = Try(s.toInt).toOption
-
-  def stringField(form: UrlForm, field: String): Either[String, String] =
-    form.getFirst(field).toRight(s"Missing field '$field'")
-
-  def intField(form: UrlForm, field: String,
-               min: Int = Int.MinValue, max: Int = Int.MaxValue): Either[String, Int] =
-    form.getFirst(field)
-      .flatMap(stringToInt)
-      .toRight(s"Invalid format for field '$field'")
-      .right.flatMap { i =>
-        if (i < min) Left(s"Number $i should be greater or equal than $min")
-        else Right(i)
-      }
-      .right.flatMap { i =>
-        if (i > max) Left(s"Number $i should be less or equal than $max")
-        else Right(i)
-      }
-}
-
 object Game {
   import FormParsing._
 
-  private val trailingNumber = "\\d+$".r
-  private val minCardNumber = 1
-  private val maxCardNumber = 52
+  def fromForm(form: UrlForm): Either[String, GameInput] = {
+    import CardGradeInput.{fromForm => cardFromForm}
 
-  def fromForm(form: UrlForm): Either[String, GameInput] =
     for {
       owner <- stringField(form, "owner").right
       email <- stringField(form, "email").right
@@ -47,49 +24,23 @@ object Game {
       openQuestion <- stringField(form, "openQuestion").right
       rating <- intField(form, "rating", min = 1, max = 5).right
     } yield GameInput(owner, email, strong, weak, important, hard, inspiring, tedious, topaasia, openQuestion, rating)
+  }
 
-  def fromInput(game: GameInput, createdAt: LocalDateTime, cards: Card.Collection): Game = {
-    def card(code: Card.Code): Card =
-      cards.getOrElse(code, Card(code, code, code))
-
-    def cardGrade(ci: CardGradeInput): CardGrade = {
-      val c = card(ci.code)
-      CardGrade(code = ci.code, subject = c.subject, sentence = c.sentence, grade = ci.grade)
-    }
-
+  def fromInput(game: GameInput, createdAt: LocalDateTime, cards: Card.Collection): Game =
     Game(
       owner = game.owner,
       email = game.email,
-      strong = cardGrade(game.strong),
-      weak = cardGrade(game.weak),
-      important = cardGrade(game.important),
-      hard = cardGrade(game.hard),
-      tedious = cardGrade(game.tedious),
-      inspiring = cardGrade(game.inspiring),
-      topaasia = card(game.topaasia),
+      strong = cards.graded(game.strong),
+      weak = cards.graded(game.weak),
+      important = cards.graded(game.important),
+      hard = cards.graded(game.hard),
+      tedious = cards.graded(game.tedious),
+      inspiring = cards.graded(game.inspiring),
+      topaasia = cards(game.topaasia),
       topaasiaAnswer = game.topaasiaAnswer,
       rating = game.rating,
       createdAt = createdAt
     )
-  }
-
-  private def cardFromForm(form: UrlForm, level: String) = for {
-    card <- cardName(form, s"${level}Card").right
-    grade <- intField(form, s"${level}Num", min = 1, max = 4).right
-  } yield CardGradeInput(card, grade)
-
-  private def cardName(form: UrlForm, field: String) =
-    stringField(form, field)
-      .right.map(_.toUpperCase)
-      .right.flatMap(filterCardNumber)
-
-  private def filterCardNumber(s: String): Either[String, String] =
-    trailingNumber
-      .findFirstIn(s)
-      .flatMap(stringToInt)
-      .filter(n => n >= minCardNumber && n <= maxCardNumber)
-      .map(_ => s)
-      .toRight(s"Card name '$s' should end with a number between $minCardNumber-$maxCardNumber.")
 }
 
 final case class GameInput(
@@ -110,8 +61,6 @@ final case class GameInput(
   val codes: Set[Card.Code] = cardGrades.map(_.code).toSet
 }
 
-final case class CardGradeInput(code: Card.Code, grade: Int)
-
 final case class Game(
   owner: String,
   email: String,
@@ -126,7 +75,3 @@ final case class Game(
   rating: Int,
   createdAt: LocalDateTime
 )
-
-final case class CardGrade(code: Card.Code, subject: String, sentence: String, grade: Int) {
-  def render: String = s"$subject ($grade)"
-}
